Move tagger and getGPUTags
