+++ comp/process/gpusubscriber/component.go
@@ -8,6 +8,5 @@ package gpusubscriber
 
 // team: container-intake
 
-//nolint:revive // TODO(PROC) Fix revive linter
 type Component interface {
 }
