+++ comp/process/gpusubscriber/gpusubscriberimpl/subscriber.go
@@ -17,6 +17,7 @@ import (
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	"github.com/DataDog/datadog-agent/comp/process/gpusubscriber"
 	procSubscribers "github.com/DataDog/datadog-agent/pkg/process/subscribers"
+	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
@@ -32,6 +33,7 @@ type gpusubscriberimpl struct {
 
 type dependencies struct {
 	fx.In
+	Lc fx.Lifecycle
 
 	Sysconfig sysprobeconfig.Component
 	Config    config.Component
@@ -46,9 +48,21 @@ type result struct {
 }
 
 func newGpuSubscriber(deps dependencies) result {
-	gpuSubscriber := procSubscribers.NewGPUSubscriber(deps.WMeta)
+	gpuSubscriber := procSubscribers.NewGPUSubscriber(deps.WMeta, deps.Tagger)
+	gpuSubComponent := gpusubscriberimpl{
+		gpuSubscriber: gpuSubscriber,
+	}
+
+	// TODO: only run in core agent (not process agent), when gpu & process check is enabled?
+	if flavor.GetFlavor() != flavor.ProcessAgent {
+		deps.Lc.Append(fx.Hook{
+			OnStart: gpuSubComponent.Run,
+			OnStop:  gpuSubComponent.stop,
+		})
+	}
+
 	return result{
-		Component: gpuSubscriber,
+		Component: &gpuSubComponent,
 	}
 }
 
@@ -57,7 +71,7 @@ func (g *gpusubscriberimpl) Run(context.Context) error {
 	return nil
 }
 
-func (g *gpusubscriberimpl) Stop(context.Context) error {
+func (g *gpusubscriberimpl) stop(context.Context) error {
 	g.gpuSubscriber.Stop()
 	return nil
 }
